refactor: trying to organize lookup code further
